Make run.sh work on Windows.
Create the VM if not present.

Use docker-machine on both MacOSX and Windows.  'command' is not available on
Windows.  'docker-machine' does not currently come with the standard Linux
package, but it may in the future.

Do not use -v to mount to local directory -- this does not work on Windows.